+++ blog/src/main/java/com/masai/service/PostServiceImpl.java
@@ -4,11 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.masai.exception.PostNotFoundException;
 import com.masai.model.Post;
 import com.masai.model.PostDTO;
+import com.masai.repository.CommentsDao;
 import com.masai.repository.PostDao;
 
 @Service
@@ -16,10 +20,14 @@ public class PostServiceImpl implements PostService{
 
     @Autowired
     private PostDao postDao;
+    
+    @Autowired
+    private CommentsDao commentsDao;
 
     @Override
     public List<PostDTO> getAllPosts() {
         List<Post> list = postDao.findAll();
+        if(list.size()==0) throw new PostNotFoundException("no post found");
         List<PostDTO> postDTOList = new ArrayList<>();
         for(int i=0; i<list.size(); i++)
         {
@@ -33,6 +41,9 @@ public class PostServiceImpl implements PostService{
     @Override
     public PostDTO getPostById(int postId) {
         Optional<Post> opt = postDao.findById(postId);
+        if(!opt.isPresent()) {
+        	throw new PostNotFoundException("post with this id not found");
+        }
         PostDTO post = new PostDTO(opt.get().getContent(),opt.get().getTitle());
         return post;
     }
@@ -45,19 +56,31 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public PostDTO updatePostById(int postId, Post post) {
+    public PostDTO updatePostById(int postId, PostDTO postdto) {
         Optional<Post> opt = postDao.findById(postId);
         if(opt.get() != null)
         {
-            postDao.save(post);
-            PostDTO post2 = new PostDTO(post.getContent(),post.getTitle());
-            return post2;
+//        	post.setPostId(postId);
+//        	post.setList(opt.get().getList());
+//            postDao.save(post);
+//            PostDTO post2 = new PostDTO(post.getContent(),post.getTitle());
+//            return post2;
+        	opt.get().setContent(postdto.getContent());
+        	opt.get().setTitle(postdto.getTitle());
+        	postDao.save(opt.get());
+        	return postdto;
         }
-        return null;
+        throw new PostNotFoundException("post with this id not found");
     }
 
+//    @Transactional
     @Override
     public String deletePostById(int postId) {
+    	Optional<Post> opt = postDao.findById(postId);
+        if(!opt.isPresent()) {
+        	throw new PostNotFoundException("post with this id not found");
+        }
+//        commentsDao.deleteCommentsByPostId(postId);
         postDao.deleteById(postId);
         return "post successfully deleted";
     }
